redirector/scripts/redirector_ip/requirements.txt
boto3
requests
dnspython